(MAINT) Add initial structure.
This commit adds two new top-level folders to the project: one for the
documentation and one for the source code.

The `Documentation` folder is minimal and only includes a readme.

The `Source` folder is more substantial, laying down files and folders
to support the proposed structure and design of this module, with readme
files to explain the general purpose.

The `Source` folder also contains the minimal default manifest for the
PowerShell module and a blank script file.

The `Classes` and `Enums` folders in `Public` and `Private` also contain
a `LoadOrder.jsonc` file, which points at how we might resolve the load
order for those items during testing and development.

The `.gitkeep` files exist only until other files can be placed in those
folders, at which time the `gitkeep` files will no longer be required.